Fix Pyton and C implementations
Both implementations now return the same values, and one can be used to
verify signatures that were created by the other.

+++ C/adrs.c
@@ -3,18 +3,18 @@
 #include "adrs.h"
 
 // Hilfsfunktion: Wandelt eine Ganzzahl in ein Byte-Array um
-void toByte(uint32_t x, uint32_t n, uint8_t *S)
+void toByte(uint64_t x, uint64_t n, uint8_t *S)
 {
-    for (uint32_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         S[n - 1 - i] = x % 256;
         x >>= 8;
     }
 }
 
 // Hilfsfunktion: Wandelt ein Byte-Array in eine Ganzzahl um
-uint32_t toInt(uint8_t *X, uint32_t n) {
-    uint32_t total = 0;
-    for (uint32_t i = 0; i < n; i++) {
+uint64_t toInt(uint8_t *X, uint64_t n) {
+    uint64_t total = 0;
+    for (uint64_t i = 0; i < n; i++) {
         total = 256 * total + X[i];
     }
     return total;
@@ -25,43 +25,43 @@ void initADRS(ADRS *adrs) {
     memset(adrs->adrs, 0, ADRS_SIZE);
 }
 
-void setLayerAddress(ADRS *adrs, uint32_t l) {
+void setLayerAddress(ADRS *adrs, uint64_t l) {
     toByte(l, 4, adrs->adrs);
 }
 
-void setTreeAddress(ADRS *adrs, uint32_t t) {
+void setTreeAddress(ADRS *adrs, uint64_t t) {
     toByte(t, 12, adrs->adrs + 4);
 }
 
-void setTypeAndClear(ADRS *adrs, uint32_t Y) {
+void setTypeAndClear(ADRS *adrs, uint64_t Y) {
     toByte(Y, 4, adrs->adrs + 16);
     memset(adrs->adrs + 20, 0, 12);
 }
 
-void setKeyPairAddress(ADRS *adrs, uint32_t i) {
+void setKeyPairAddress(ADRS *adrs, uint64_t i) {
     toByte(i, 4, adrs->adrs + 20);
 }
 
-void setChainAddress(ADRS *adrs, uint32_t i) {
+void setChainAddress(ADRS *adrs, uint64_t i) {
     toByte(i, 4, adrs->adrs + 24);
 }
 
-void setTreeHeight(ADRS *adrs, uint32_t i) {
+void setTreeHeight(ADRS *adrs, uint64_t i) {
     toByte(i, 4, adrs->adrs + 24);
 }
 
-void setHashAddress(ADRS *adrs, uint32_t i) {
+void setHashAddress(ADRS *adrs, uint64_t i) {
     toByte(i, 4, adrs->adrs + 28);
 }
 
-void setTreeIndex(ADRS *adrs, uint32_t i) {
+void setTreeIndex(ADRS *adrs, uint64_t i) {
     toByte(i, 4, adrs->adrs + 28);
 }
 
-uint32_t getKeyPairAddress(ADRS *adrs) {
+uint64_t getKeyPairAddress(ADRS *adrs) {
     return toInt(adrs->adrs + 20, 4);
 }
 
-uint32_t getTreeIndex(ADRS *adrs) {
+uint64_t getTreeIndex(ADRS *adrs) {
     return toInt(adrs->adrs + 28, 4);
 }
