+++ C/adrs.h
@@ -5,31 +5,31 @@
 #define ADRS_SIZE 32
 
 typedef struct {
-    uint8_t adrs[ADRS_SIZE]; // 32 Byte Address
+    uint8_t adrs[ADRS_SIZE];
 } ADRS;
 
-void toByte(uint32_t x, uint32_t n, uint8_t *S);
+void toByte(uint64_t x, uint64_t n, uint8_t *S);
 
-uint32_t toInt(uint8_t *X, uint32_t n);
+uint64_t toInt(uint8_t *X, uint64_t n);
 
 void initADRS(ADRS *adrs);
 
-void setLayerAddress(ADRS *adrs, uint32_t l);
+void setLayerAddress(ADRS *adrs, uint64_t l);
 
-void setTreeAddress(ADRS *adrs, uint32_t t);
+void setTreeAddress(ADRS *adrs, uint64_t t);
 
-void setTypeAndClear(ADRS *adrs, uint32_t Y);
+void setTypeAndClear(ADRS *adrs, uint64_t Y);
 
-void setKeyPairAddress(ADRS *adrs, uint32_t i);
+void setKeyPairAddress(ADRS *adrs, uint64_t i);
 
-void setChainAddress(ADRS *adrs, uint32_t i);
+void setChainAddress(ADRS *adrs, uint64_t i);
 
-void setTreeHeight(ADRS *adrs, uint32_t i);
+void setTreeHeight(ADRS *adrs, uint64_t i);
 
-void setHashAddress(ADRS *adrs, uint32_t i);
+void setHashAddress(ADRS *adrs, uint64_t i);
 
-void setTreeIndex(ADRS *adrs, uint32_t i);
+void setTreeIndex(ADRS *adrs, uint64_t i);
 
-uint32_t getKeyPairAddress(ADRS *adrs);
+uint64_t getKeyPairAddress(ADRS *adrs);
 
-uint32_t getTreeIndex(ADRS *adrs);
+uint64_t getTreeIndex(ADRS *adrs);
