+++ C/fors.h
@@ -3,9 +3,9 @@
 #include "params.h"
 #include "adrs.h"
 
-void fors_skGen(Parameters *prm, const uint8_t *sk_seed, const uint8_t *pk_seed, ADRS adrs, uint32_t idx, uint8_t *buffer);
+void fors_skGen(Parameters *prm, const uint8_t *sk_seed, const uint8_t *pk_seed, ADRS adrs, uint64_t idx, uint8_t *buffer);
 
-void fors_node(Parameters *prm, const uint8_t *sk_seed, uint32_t i, uint32_t z, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer);
+void fors_node(Parameters *prm, const uint8_t *sk_seed, uint64_t i, uint64_t z, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer);
 
 void fors_sign(Parameters *prm, const uint8_t *md, const uint8_t *sk_seed, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer);
