+++ C/xmss.c
@@ -5,7 +5,7 @@
 #include "xmss.h"
 
 // algorithm 9
-void xmss_node(Parameters *prm, const uint8_t* sk_seed, uint32_t i, uint32_t z, const uint8_t* pk_seed, ADRS adrs, uint8_t *buffer)
+void xmss_node(Parameters *prm, const uint8_t* sk_seed, uint64_t i, uint64_t z, const uint8_t* pk_seed, ADRS adrs, uint8_t *buffer)
 {
     uint8_t node[prm->n];
 
@@ -34,11 +34,11 @@ void xmss_node(Parameters *prm, const uint8_t* sk_seed, uint32_t i, uint32_t z, 
 }
 
 // algorithm 10
-void xmss_sign(Parameters *prm, const uint8_t *M, const uint8_t *sk_seed, uint32_t idx, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer)
+void xmss_sign(Parameters *prm, const uint8_t *M, const uint8_t *sk_seed, uint64_t idx, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer)
 {
     uint8_t AUTH[prm->h_ * prm->n];
     for (uint32_t j = 0; j < prm->h_; j++) {
-        uint32_t k = (uint32_t) floor(idx / pow(2, j)) ^ 1;
+        uint64_t k = (uint64_t) floor(idx / pow(2, j)) ^ 1;
         // Alternative: uint32_t k = (idx >> j) ^ 1;
         xmss_node(prm, sk_seed, k, j, pk_seed, adrs, AUTH + j * prm->n);
     }
@@ -53,7 +53,7 @@ void xmss_sign(Parameters *prm, const uint8_t *M, const uint8_t *sk_seed, uint32
 }
 
 // algorithm 11
-void xmss_pkFromSig(Parameters *prm, uint32_t idx, const uint8_t *sig_xmss, const uint8_t *M, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer)
+void xmss_pkFromSig(Parameters *prm, uint64_t idx, const uint8_t *sig_xmss, const uint8_t *M, const uint8_t *pk_seed, ADRS adrs, uint8_t *buffer)
 {
     uint8_t node_0[prm->n];
     uint8_t node_1[prm->n];
@@ -74,7 +74,7 @@ void xmss_pkFromSig(Parameters *prm, uint32_t idx, const uint8_t *sig_xmss, cons
     uint8_t combined[2 * prm->n];
     for (uint32_t k = 0; k < prm->h_; k++) {
         setTreeHeight(&adrs, k + 1);
-        if ((int32_t) floor(idx / pow(2, k)) % 2 == 0) {
+        if ((uint64_t) floor(idx / pow(2, k)) % 2 == 0) {
             setTreeIndex(&adrs, getTreeIndex(&adrs) / 2);
             memcpy(combined, node_0, prm->n);
             memcpy(combined + prm->n, AUTH + k * prm->n, prm->n);
